Improved Depends/Imports/Suggests as required by CRAN these days:
- flexmix is now only in Imports but not Suggests and hence only
requireNamespace("flexmix") can be used but not require("flexmix").
(For the latter, it had to be in Suggests. But we need it in Imports as
well and CRAN doesn't want to have the same package in Imports and
Suggests.)

- Consequently, library("flexmix") is needed when calling FLXPmultinom()
on the manual pages now (because this is done by the user and not bu
betareg).

- A few further enhancements related to this and updated example output.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/betareg@172 edb9625f-4e0d-4859-8d74-9fd3b1da38cb